NN throwing an error
What is supposed to happen is 3 individual Antagonist NN are trained and the weights of each are saved into a vector.  Then the Overlord Antagonist chooses which of the 3 individual NN to choose from for a given timestep.  It's not getting past the 1 generation out of 50 generations for the 1st individual Antagonist NN.